utils: moved objectMap, expectDeepSubsetMatch, objectFilterRecursiveToMatchStructure + added testUtils
