+++ src/utils/object.js
@@ -78,3 +78,51 @@ const objr = (fn, {o=null, n=null} = {})=> {
 }
 
 export const stupidIterativeObjectDependencyResolve = objr
+
+
+// eg. obj = {a: 2, b: 3, c: {some: 'more', even: 'more'}, g: {bla: 5}}
+// 	structure = {a:1, c: {some:8}, g: '..', y: 9}
+// (obj, structure) -> {a: 2, c: {some: 'more'}, g: {bla: 5}, y: undefined}
+// TODO: rename to subset..? + fix array to be subset based; not key/index-based
+export const objectFilterRecursiveToMatchStructure = (obj, structure, {
+	taken = new Set(),
+	takenSkip = false, takenReturnStructure = false,
+} = {})=>
+		typeof obj !== 'object' || typeof structure !== 'object'
+		|| obj===null || structure===null
+		|| obj===structure ? obj
+	: taken.has(structure)? takenSkip
+		? void 0
+		: takenReturnStructure? structure: obj
+	: (taken.add(structure), Object.keys(structure).reduce((o, k)=> (o[k] =
+		objectFilterRecursiveToMatchStructure(obj[k], structure[k], {
+			taken, takenSkip, takenReturnStructure}), o), {}))
+
+export const arrayAppend = (target, ...xss)=> (xss.forEach(xs=> target.push(...xs)), target)
+
+export const objectMap = fn=> obj=> Object.keys(obj)
+	.reduce((o, k)=> (o[k] = fn(obj[k], k), o), {})
+
+export const objectMapRecursive = (obj, fn, {
+	// filter = ({})
+	taken = [], takenMapCache = [],
+	key = null,
+} = {})=> typeof obj !== 'object' || obj===null
+	? fn(obj, key, {recurse: null})
+	: taken.indexOf(obj)>=0
+		? takenMapCache[taken.indexOf(obj)]
+		: (taken.push(obj), Array.isArray(obj)
+			// TODO: also do object properties on arrays (non integer keys)
+			? arrayAppend(takenMapCache[takenMapCache.length] = [],
+				obj.map((v, k)=> fn(v, k, {
+					recurse: ()=> objectMapRecursive(v, fn, {
+						taken, takenMapCache, key: k,
+					}),
+				})))
+			: Object.assign(takenMapCache[takenMapCache.length] = {},
+				objectMap((v, k)=> fn(v, k, {
+					recurse: ()=> objectMapRecursive(v, fn, {
+						taken, takenMapCache, key: k,
+					}),
+				}))(obj))
+		)
